Centered icons in statusbar, some fixes.
git-svn-id: http://miranda.googlecode.com/svn/trunk@1187 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/mwclist/clc.c
@@ -56,7 +56,7 @@ void StatusUpdaterThread(HWND hwndDlg)
 	curdelay=DBGetContactSettingByte(hContact,"CList","StatusMsgAutoDelay",15000);
 	if (curdelay<5000) curdelay=5000;
 	
-	if ((GetTickCount()-lastcheck)>curdelay)
+	if ((int)(GetTickCount()-lastcheck)>curdelay)
 	{
 		lastcheck=GetTickCount();
 		if (DBGetContactSettingByte(hContact,"CList","StatusMsgAuto",0)) {
